NXDRIVE-45: WiP: Fix the remote modification

+++ nuxeo-drive-client/nxdrive/engine/watcher/local_watcher.py
@@ -165,9 +165,15 @@ class LocalWatcher(Worker):
                     doc_pair.remote_parent_ref == remote_parent_ref):
                 log.debug("Detect move for %s (%r)", local_info.name, doc_pair)
                 doc_pair.local_state = 'moved'
+            elif doc_pair.local_state == 'moved':
+                # The pair was moved but it has been canceled manually
+                doc_pair.local_state = 'synchronized'
             self._dao.update_local_state(doc_pair, local_info)
             return
         if evt.event_type == 'deleted':
+            if self.client.exists(doc_pair.local_path):
+                # This happens on update dont do anything
+                return
             doc_pair.update_state('deleted', doc_pair.remote_state)
             if doc_pair.remote_state == 'unknown':
                 self._dao.remove_state(doc_pair)
